fix: filter by activity done

+++ src/Pipedrive.net/Models/Request/ActivityFilters.cs
@@ -52,7 +52,7 @@ namespace Pipedrive
                 }
                 if (Done.HasValue)
                 {
-                    d.Add("done", Done.Value.ToString());
+                    d.Add("done", ((int)Done.Value).ToString());
                 }
                 return d;
             }
